requirements.txt
# Core
websockets>=16.0.0
python-dotenv>=1.2.1
diff-match-patch>=20241021

# Just for unit tests
# pytest

# Gemini Live provider + optional LLM benchmark metric
google-genai>=1.64.0

# Providers
#
# Cartesia
# We do not use their library, we directly stream to webscokcet.
#
# Eleven Labs
# We do not use their library, we directly stream to webscokcet.
# elevenlabs>=2.31.0
#
# Google
google-cloud-speech>=2.36.0
#
# Speechmatics
# We do not use their library, we directly stream to webscokcet.
